Moved regulariz to functions.py + added tests
