+++ src/functions.py
@@ -185,10 +185,33 @@ def binary_class_accuracy(predicted, target):
     return np.array([0])
 
 
-BinClassAcc = Function(binary_class_accuracy, 'class_acc')
-metrics = {
-    'bin_class_acc': BinClassAcc
-}
+""" Regularizations """
+
+
+def lasso_l1(w, lambd):
+    return lambd * np.sum(np.abs(w))
+
+
+def lasso_l1_deriv(w, lambd):
+    res = np.zeros(w.shape)
+    for i in range((w.shape[0])):
+        for j in range(w.shape[1]):
+            if w[i][j] < 0:
+                res[i][j] = -lambd
+            elif w[i][j] > 0:
+                res[i][j] = lambd
+            else:
+                res[i][j] = 0
+    return res
+
+
+def ridge_l2(w, lambd):
+    return lambd * np.sum(np.square(w))
+
+
+def ridge_l2_deriv(w, lambd):
+    return 2 * lambd * w
+
 
 ReLU = DerivableFunction(relu, relu_deriv, 'ReLU')
 LeakyReLU = DerivableFunction(leaky_relu, leaky_relu_deriv, 'LeakyReLU')
@@ -206,3 +229,14 @@ losses = {
     'squared': SquaredLoss,
 }
 
+l2_regularization = DerivableFunction(ridge_l2, ridge_l2_deriv, 'l2')
+l1_regularization = DerivableFunction(lasso_l1, lasso_l1_deriv, 'l1')
+regs = {
+    'l2': l2_regularization,
+    'l1': l1_regularization
+}
+
+BinClassAcc = Function(binary_class_accuracy, 'class_acc')
+metrics = {
+    'bin_class_acc': BinClassAcc
+}
